trunk 	gs sync to l2j 966
- fix for being dismissed from clan while in clan hall
- Silent Move behavior fix (by Venom) 
- Updated the abnormals effects. Added stealth one.
  Since i didnt hardcoded it u will need to add abnormal="stealth" to the effect in order to get it.

git-svn-id: http://svn.l2jfree.com/l2jfree@1669 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/L2_GameServer_IL/src/main/java/net/sf/l2j/gameserver/handler/admincommandhandlers/AdminEffects.java
@@ -126,9 +126,9 @@ public class AdminEffects implements IAdminCommandHandler
 	    			player = (L2Character)target;
 	    			
 	    			if (type.equals("1"))
-	    			   player.startAbnormalEffect((short)0x0400);
+	    				player.startAbnormalEffect(0x0400);
 	    			else
-	    				player.startAbnormalEffect((short)0x0800);
+	    				player.startAbnormalEffect(0x0800);
 	            	player.setIsParalyzed(true);
 	            	
 	            	StopMove sm = new StopMove(player);
@@ -164,7 +164,7 @@ public class AdminEffects implements IAdminCommandHandler
                {
             	   if (!player.isGM())
             	   {
-                	   player.startAbnormalEffect((short)0x0400);
+            		   player.startAbnormalEffect(0x0400);
                 	   player.setIsParalyzed(true);
                 	   
                 	   StopMove sm = new StopMove(player);
@@ -183,7 +183,7 @@ public class AdminEffects implements IAdminCommandHandler
            {
                for (L2PcInstance player : activeChar.getKnownList().getKnownPlayers().values())
                {
-            	   player.stopAbnormalEffect((short)0x0400);
+            	   player.stopAbnormalEffect(0x0400);
             	   player.setIsParalyzed(false);
                }
            }
@@ -229,7 +229,7 @@ public class AdminEffects implements IAdminCommandHandler
                L2Character player = null;
                 if (target instanceof L2Character) {
                   player = (L2Character)target;
-                  player.startAbnormalEffect((short)0x2000);
+                  player.startAbnormalEffect(0x2000);
                 }
            }
            catch (Exception e)
